20042025-1122

+++ games/stonecraft_game/mods/mesecons/mesecons_mvps/init.lua
@@ -1,13 +1,3 @@
---[[
-
-2017-01-06 modified by MrCerealGuy <mrcerealguy@gmx.de>
-	exit if mod is deactivated
-
---]]
-
--- pipeworks needs mesecon_mvps mod
-if core.skip_mod("mesecons") and core.skip_mod("pipeworks") then return end
-
 --register stoppers for movestones/pistons
 
 mesecon.mvps_stoppers = {}
@@ -56,88 +46,73 @@ local function on_mvps_move(moved_nodes)
 	end
 end
 
-function mesecon.mvps_process_stack(stack)
-	-- update mesecons for placed nodes ( has to be done after all nodes have been added )
-	for _, n in ipairs(stack) do
-		mesecon.on_placenode(n.pos, minetest.get_node(n.pos))
-	end
+function mesecon.mvps_process_stack()
+	-- This function is kept for compatibility.
+	-- It used to call on_placenode on moved nodes, but that is now done automatically.
 end
 
 -- tests if the node can be pushed into, e.g. air, water, grass
 local function node_replaceable(name)
-	if minetest.registered_nodes[name] then
-		return minetest.registered_nodes[name].buildable_to or false
+	local nodedef = minetest.registered_nodes[name]
+
+	-- everything that can be an mvps stopper (unknown nodes and nodes in the
+	-- mvps_stoppers table) must not be replacable
+	-- Note: ignore (a stopper) is buildable_to, but we do not want to push into it
+	if not nodedef or mesecon.mvps_stoppers[name] then
+		return false
 	end
 
-	return false
+	return nodedef.buildable_to or false
 end
 
 function mesecon.mvps_get_stack(pos, dir, maximum, all_pull_sticky)
 	-- determine the number of nodes to be pushed
 	local nodes = {}
-	local frontiers = {pos}
+	local pos_set = {}
+	local frontiers = mesecon.fifo_queue.new()
+	frontiers:add(vector.new(pos))
 
-	while #frontiers > 0 do
-		local np = frontiers[1]
-		local nn = minetest.get_node(np)
-
-		if not node_replaceable(nn.name) then
+	for np in frontiers:iter() do
+		local np_hash = minetest.hash_node_position(np)
+		local nn = not pos_set[np_hash] and minetest.get_node(np)
+		if nn and not node_replaceable(nn.name) then
+			pos_set[np_hash] = true
 			table.insert(nodes, {node = nn, pos = np})
 			if #nodes > maximum then return nil end
 
-			-- add connected nodes to frontiers, connected is a vector list
-			-- the vectors must be absolute positions
-			local connected = {}
-			if minetest.registered_nodes[nn.name]
-			and minetest.registered_nodes[nn.name].mvps_sticky then
-				connected = minetest.registered_nodes[nn.name].mvps_sticky(np, nn)
+			-- add connected nodes to frontiers
+			local nndef = minetest.registered_nodes[nn.name]
+			if nndef and nndef.mvps_sticky then
+				local connected = nndef.mvps_sticky(np, nn)
+				for _, cp in ipairs(connected) do
+					frontiers:add(cp)
+				end
 			end
 
-			table.insert(connected, vector.add(np, dir))
+			frontiers:add(vector.add(np, dir))
 
 			-- If adjacent node is sticky block and connects add that
-			-- position to the connected table
+			-- position
 			for _, r in ipairs(mesecon.rules.alldirs) do
 				local adjpos = vector.add(np, r)
 				local adjnode = minetest.get_node(adjpos)
-				if minetest.registered_nodes[adjnode.name]
-				and minetest.registered_nodes[adjnode.name].mvps_sticky then
-					local sticksto = minetest.registered_nodes[adjnode.name]
-						.mvps_sticky(adjpos, adjnode)
+				local adjdef = minetest.registered_nodes[adjnode.name]
+				if adjdef and adjdef.mvps_sticky then
+					local sticksto = adjdef.mvps_sticky(adjpos, adjnode)
 
 					-- connects to this position?
 					for _, link in ipairs(sticksto) do
 						if vector.equals(link, np) then
-							table.insert(connected, adjpos)
+							frontiers:add(adjpos)
 						end
 					end
 				end
 			end
 
 			if all_pull_sticky then
-				table.insert(connected, vector.subtract(np, dir))
-			end
-
-			-- Make sure there are no duplicates in frontiers / nodes before
-			-- adding nodes in "connected" to frontiers
-			for _, cp in ipairs(connected) do
-				local duplicate = false
-				for _, rp in ipairs(nodes) do
-					if vector.equals(cp, rp.pos) then
-						duplicate = true
-					end
-				end
-				for _, fp in ipairs(frontiers) do
-					if vector.equals(cp, fp) then
-						duplicate = true
-					end
-				end
-				if not duplicate then
-					table.insert(frontiers, cp)
-				end
+				frontiers:add(vector.subtract(np, dir))
 			end
 		end
-		table.remove(frontiers, 1)
 	end
 
 	return nodes
@@ -249,6 +224,8 @@ function mesecon.mvps_push_or_pull(pos, stackdir, movedir, maximum, all_pull_sti
 		minetest.remove_node(n.pos)
 	end
 
+	local oldstack = mesecon.tablecopy(nodes)
+
 	-- update mesecons for removed nodes ( has to be done after all nodes have been removed )
 	for _, n in ipairs(nodes) do
 		mesecon.on_dignode(n.pos, n.node)
@@ -258,6 +235,12 @@ function mesecon.mvps_push_or_pull(pos, stackdir, movedir, maximum, all_pull_sti
 	for _, n in ipairs(nodes) do
 		local np = vector.add(n.pos, movedir)
 
+		-- Turn off conductors in transit
+		local conductor = mesecon.get_conductor(n.node.name)
+		if conductor and conductor.state ~= mesecon.state.off then
+			n.node.name = conductor.offstate or conductor.states[1]
+		end
+
 		minetest.set_node(np, n.node)
 		minetest.get_meta(np):from_table(n.meta)
 		if n.node_timer then
@@ -266,7 +249,6 @@ function mesecon.mvps_push_or_pull(pos, stackdir, movedir, maximum, all_pull_sti
 	end
 
 	local moved_nodes = {}
-	local oldstack = mesecon.tablecopy(nodes)
 	for i in ipairs(nodes) do
 		moved_nodes[i] = {}
 		moved_nodes[i].oldpos = nodes[i].pos
@@ -283,7 +265,6 @@ function mesecon.mvps_push_or_pull(pos, stackdir, movedir, maximum, all_pull_sti
 end
 
 function mesecon.mvps_move_objects(pos, dir, nodestack, movefactor)
-	local objects_to_move = {}
 	local dir_k
 	local dir_l
 	for k, v in pairs(dir) do
@@ -341,11 +322,6 @@ end
 -- TODO: load blocks instead, as with wires.
 mesecon.register_mvps_stopper("ignore")
 
-mesecon.register_mvps_stopper("doors:door_steel_b_1")
-mesecon.register_mvps_stopper("doors:door_steel_t_1")
-mesecon.register_mvps_stopper("doors:door_steel_b_2")
-mesecon.register_mvps_stopper("doors:door_steel_t_2")
-mesecon.register_mvps_stopper("default:chest_locked")
 mesecon.register_on_mvps_move(mesecon.move_hot_nodes)
 mesecon.register_on_mvps_move(function(moved_nodes)
 	for i = 1, #moved_nodes do
